+++ incl/makeClusterPSOCK.R
@@ -240,13 +240,13 @@ cl <- makeClusterPSOCK(
 )
 
 
-## EXAMPLE: Two workers running in Singularity on the local machine
-## Setup of 2 Singularity workers running rocker/r-parallel
+## EXAMPLE: Two workers running via Linux container 'rocker/r-parallel' from
+## DockerHub on the local machine using Apptainer (formerly Singularity)
 cl <- makeClusterPSOCK(
   rep("localhost", times = 2L),
   ## Launch Rscript inside Linux container
   rscript = c(
-    "singularity", "exec", "docker://rocker/r-parallel",
+    "apptainer", "exec", "docker://rocker/r-parallel",
     "Rscript"
   ),
   dryrun = TRUE, quiet = TRUE
@@ -274,20 +274,20 @@ cl <- makeClusterPSOCK(
 ## EXAMPLE: One worker running in Wine for Linux on the local machine
 ## To install R for MS Windows in Wine, do something like:
 ##   winecfg  # In GUI, set 'Windows version' to 'Windows 10'
-##   wget https://cran.r-project.org/bin/windows/base/R-4.2.2-win.exe
-##   wine R-4.2.2-win.exe /SILENT
+##   wget https://cran.r-project.org/bin/windows/base/R-4.2.3-win.exe
+##   wine R-4.2.3-win.exe /SILENT
 ## Prevent packages from being installed to R's system library:
-##   chmod ugo-w "$HOME/.wine/drive_c/Program Files/R/R-4.2.2/library/"
+##   chmod ugo-w "$HOME/.wine/drive_c/Program Files/R/R-4.2.3/library/"
 ## Verify it works:
-##   wine "C:/Program Files/R/R-4.2.2/bin/x64/Rscript.exe" --version
+##   wine "C:/Program Files/R/R-4.2.3/bin/x64/Rscript.exe" --version
 cl <- makeClusterPSOCK(1L,
   rscript = c(
     ## Silence Wine warnings
     "WINEDEBUG=fixme-all",
-    ## Don't pass LC_*** environments from Linux to Wine
-    sprintf("%s=", grep("LC_", names(Sys.getenv()), value = TRUE)),
+    ## Don't pass LC_* and R_LIBS* environments from host to Wine
+    sprintf("%s=", grep("^(LC_|R_LIBS)", names(Sys.getenv()), value = TRUE)),
     "wine",
-    "C:/Program Files/R/R-4.2.2/bin/x64/Rscript.exe"
+    "C:/Program Files/R/R-4.2.3/bin/x64/Rscript.exe"
   ),
   dryrun = TRUE, quiet = TRUE
 )